Timeline changes

+++ app/Macros/Routing/Router.php
@@ -38,23 +38,27 @@ class Router
         }
 
         if (!DefaultRouter::hasMacro('employeeInResource')) {
-            DefaultRouter::macro('employeeInResource', function ($module) {
+            DefaultRouter::macro('employeeInResource', function ($module, $options = []) {
 
                 $url        = str_replace('.', '', Str::plural($module));
                 $name       = Str::singular($module);
                 $controller = Str::studly(str_replace('.', ' ', $module)) . 'Controller';
+                $except =  !empty($options['except']) ? $options['except'] : [];
 
                 DefaultRouter::group([
                     'middleware' => ['auth'],
-                ], function () use ($url, $name, $controller) {
+                ], function () use ($url, $name, $controller, $except) {
                     DefaultRouter::get($url . '/employee/{employee}', $controller . '@index' )->name($url);
                     
-                    DefaultRouter::get($url . '/create', $controller . '@create')->name($url . '.create');
+                    DefaultRouter::get($url . '/create/employee/{employee}', $controller . '@create')->name($url . '.create');
                     DefaultRouter::post($url, $controller . '@store')->name($url . '.store');
                     DefaultRouter::get($url . '/{' . $name .'}/edit', $controller . '@edit')->name($url . '.edit');
                     DefaultRouter::put($url . '/{' . $name .'}', $controller . '@update')->name($url . '.update');
                     DefaultRouter::patch($url . '/{' . $name .'}', $controller . '@update')->name($url . '.update');
-                    DefaultRouter::delete($url . '/{' .$name .'}', $controller . '@destroy')->name($url . '.destroy');
+                    if(!in_array('destroy', $except)){
+                        DefaultRouter::delete($url . '/{' .$name .'}', $controller . '@destroy')->name($url . '.destroy');
+                    }
+
                 });
             });
         }
